basic auth

+++ api/bug/bug.controller.js
@@ -21,7 +21,7 @@ export async function getBugs(req, res) {
     const bugs = await bugService.query(filterBy, sortBy, sortDir);
     res.status(200).send(bugs);
   } catch (err) {
-    res.status(404).send('Cannot get bugs');
+    res.status(404).send({ error: 'Cannot get bugs' });
   }
 }
 
@@ -35,7 +35,7 @@ export async function getBug(req, res) {
   res.cookie('visitCount', visitCount, { maxAge: 1000 * 7 });
   console.log(visitCount.length);
   if (visitCount.length > 3) {
-    res.status(429).send('Wait for a bit');
+    res.status(429).send({ message: 'Wait for a bit' });
     return;
   }
 
@@ -43,55 +43,71 @@ export async function getBug(req, res) {
     const bug = await bugService.getById(bugId);
     res.status(200).send(bug);
   } catch (err) {
-    res.status(404).send('Cannot get bug');
+    res.status(404).send({ error: 'Cannot get bug' });
   }
 }
 
 export async function removeBug(req, res) {
-  const { bugId } = req.params;
   try {
+    const { bugId } = req.params;
+    if (bug.creator !== req.user._id) {
+      return res.status(403).send({ error: 'Not authorized' });
+    }
+    const bug = await getById(bugId);
+    if (!bug) return res.status(404).send({ error: 'bug not found' });
     await bugService.remove(bugId);
-    res.status(204).send;
+    res.status(204).send({ message: `bug with id: ${bugId} was removed` });
   } catch (err) {
-    res.status(404).send(`Cannot remove bug, id:(${bugId})`);
+    res.status(400).send({ error: `Cannot remove bug, id:(${bugId})` });
   }
 }
 
 export async function updateBug(req, res) {
   const { bugId } = req.params;
-  const queryObject = req.query;
-  console.log(queryObject);
+  const { title, description, severity, labels } = req.query;
+  const queryObject = { title, description, severity, labels };
   let incomingBug = await bugService.getById(bugId);
+  if (!incomingBug) {
+    return res.status(404).send({
+      error: `the bug that was requested for updating does not exist id: ${bugId}`,
+    });
+  }
   incomingBug = {
     ...incomingBug,
     ...queryObject,
   };
 
   try {
-    // note the method acts like PATCH when the object exists and not like PUT
-    const [_, bugExists] = await bugService.save(incomingBug);
-    if (bugExists)
-      res.status(204).send(); // signal the the resource exists and was updated
-    else res.status(201).send(); // signal the the resource did not exist and the PUT request created it
+    // note the method acts like PATCH, and not like PUT
+    await bugService.save(incomingBug);
+    res.status(204).send(); // signal the the resource exists and was updated
   } catch (err) {
-    res.status(400).send('Cannot save bugs');
+    res.status(400).send({ error: 'Cannot save bugs' });
   }
 }
 
 export async function addBug(req, res) {
   const {
-    title = '',
-    severity = '',
-    description = '',
-    createdAt = '',
-    labels = [],
-  } = req.params;
-  const bugToSave = { title, severity, description, createdAt, labels };
+    title,
+    severity,
+    description,
+    createdAt = new Date(),
+    labels,
+  } = req.body;
+  const creator = req.user._id;
+  const bugToSave = {
+    title,
+    severity,
+    description,
+    createdAt,
+    labels,
+    creator,
+  };
   try {
     const savedBug = await bugService.save(bugToSave);
     res.status(201).send(savedBug);
   } catch (err) {
     console.log(err);
-    res.status(400).send('Cannot save bug');
+    res.status(400).send({ error: 'Cannot save bug' });
   }
 }
